backend/requirements.txt
django >=5.0, <5.1
djangorestframework
pyyaml
requests
django-cors-headers


        